revert to use list instead of cfifo for dev's packet queue.
We chose to use cfifo for it's lock-free nature, but during development
and testing cycles we found it has multiple problems.

The most serious one is that cfifo has size limit, and that the buffer
can overrun when reader is not as fast as the writer.  when the content
of the cfifo is pointers to dynamically allocated buffers, which is the
case for dev's packets, when the overrun happens, all the pointers in
the cfifo are lost and will have no chance to be freed -- memory leak.
Linked list has no size limit, and readers and writers never collide if
properly synced.

Second, cfifo must be pre-allocated for each dev, which is a memory
overhead (32*4=128 bytes).
Linked list still has an overhead for the list_head struct, but is
considered costless compared to it's usage.

Third, though cfifo is lock-free, we are actually still using locking in
the packet mix code, because we need to protect reading and writing
over other variables.