use the ld linker script based initcall.
This eliminates the use of the gcc's __attribute__((constructor))
feature, but introduces use of __attribute__((section)). By doing
so we can avoid anything goes before main(), and (possibly) free
the code and data we used during init stage, which is not
implemented yet (well, that is not a lot of memory).

The code is almost all copied from the linux kernel 2.4. A major
difference is that we use a much simpler linker script. We do not
alloc the .initcall.init section in the script, but let the linker
do this automatically. The linker will practically place it
immediately after the .data section. All we have done is finding
out the start and end address of the section.