use cfifo for dev's packet queue instead of list.
this is a partial revert of ec29a47514be69d6b25fffe1bf54a2cbd779ea52.

we originally used fifo, but changed to list. The reason is listed in
the commit.

then we found that locking against list in the mixing context is too
expensive, so we are moving back to fifo.

the goal is removal of the locking.