mix: remove mix_stat and it's lock.
the idea of mix_stat and mix_mask was reducing the load of checking if
all queues are not empty. but it requires careful locking, which is
expensive in the mixing context.

checking each of the queues should not be a problem, as it's just issued
between sleeps.

the mixbit of device is kept there, but it has been virtually degraded
to be a bool value, indicating whether it's outstanding.